nenzfr: Use Bezier curve instead of splines to define nozzle contour
We now standardise the way that we specify the nozzle contours (at
least for T4 nozzles). Instead of using splines as we did previously
(which gives us issues when gridding), we now use a Bezier curve to
define expansion region of the nozzle contour. The control points of
Bezier curves defining the Mach 4, 6, 8 and 10 nozzle contours are in
nenzfr_data_files/Bezier-control-pts-t4-m*.data. The old
contour-t4-m*.data files are still kept in the same folder, but should
not be used for nenzfr calculations.

For the Mach 4, 6 and 10 nozzles, the Bezier curves need to be better
fitted - changes will be done in my next commit.

For the Mach 10 nozzle, the Bezier curve has been fitted using points
that start at x = 0.01056m. This allows us to standardise the
nozzle.input.template - so now we build a constant-area throat region
for every nozzle. Luke, I hope this does not break any of work that
you have been doing on nenzfr.

+++ app/nenzfr/nenzfr.py
@@ -92,7 +92,7 @@ def main():
                   help=("number of axial blocks for the divergence section (nozzle_blk)"))
     op.add_option('--nbj', dest='nbj', type='int', default=1,
                   help=("number of radial blocks"))
-    op.add_option('--bx', dest='bx', type='float', default=1.05,
+    op.add_option('--bx', dest='bx', type='float', default=1.10,
                   help=("clustering in the axial direction"))
     op.add_option('--by', dest='by', type='float', default=1.002,
                   help=("clustering in the radial direction"))
@@ -144,7 +144,8 @@ def main():
         return -2
     #
     # Get the nozzle contour file into the current work directory.
-    run_command('cp '+E3BIN+'/nenzfr_data_files/'+opt.contourFileName+' .')
+    if not os.path.exists(opt.contourFileName):
+        run_command('cp '+E3BIN+'/nenzfr_data_files/'+opt.contourFileName+' .')
     # Set up the equilibrium gas-model file as a look-up table.
     if opt.chemModel in ['eq',]:
         if opt.gasName in ['n2']:
@@ -152,7 +153,6 @@ def main():
         else:
             eqGasModelFile = 'cea-lut-'+opt.gasName+'.lua.gz'
         if not os.path.exists(eqGasModelFile):
-            #run_command('build-cea-lut --case='+opt.gasName+' --extrapolate')
             run_command('build-cea-lut.py --gas='+opt.gasName)
         gmodelFile = eqGasModelFile
     else:
